Refactor with the FileClassifier
